app/requirements.txt
asgiref==3.7.2
Django==4.2
django-neomodel==0.1.1
factory-boy==3.3.0
Faker==23.2.1
neo4j==5.12.0
neobolt==1.7.17
neomodel==5.1.2
psycopg==3.1.18
psycopg-binary==3.1.18
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
